lab-03/requirements.txt
PyQt5p
requests
Flask>=2.3.2
rsa>=4.9
ecdsa